pyproject.toml
[tool.poetry]
name = "acled-classification"
version = "0.1.0"
description = ""
authors = ["Friederike L. Bauer <52316159+friederikelbauer@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "acled_classification"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
hiclass = "^4.2.3"
pandas = "^1.5.1"
matplotlib = "^3.6.2"
sklearn = "^0.0.post1"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
gensim = "^4.2.0"
seaborn = "^0.12.1"
plotly = "^5.11.0"
datetime = "^4.7"
pybind11 = "^2.10.1"
fasttext-wheel = "^0.9.2"
tensorflow = "^2.11.0"
transformers = "^4.24.0"
scienceplots = "^2.0.1"
torch = "^1.13.1"
sentence-transformers = "^2.2.2"
numpy = "^1.24.2"
logging = "^0.4.9.6"
pathlib = "^1.0.1"
tensorflow-hub = "^0.12.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}
ipykernel = "^6.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
